+++ src/hdmf/common/hierarchicaltable.py
@@ -57,6 +57,12 @@ def to_hierarchical_dataframe(dynamic_table):
         rows = hcol.get(slice(None), index=False, df=True)
     else:
         rows = [hcol[i:(i+1)] for i in range(len(hcol))]
+    # Retrieve the columns we need to iterate over from our input table. For AlignedDynamicTable we need to
+    # use the get_colnames function instead of the colnames property to ensure we get all columns not just
+    # the columns from the main table
+    dynamic_table_colnames = (dynamic_table.get_colnames(include_category_tables=True, ignore_category_ids=False)
+                              if isinstance(dynamic_table, AlignedDynamicTable)
+                              else dynamic_table.colnames)
 
     # Case 1:  Our DynamicTableRegion column points to a DynamicTable that itself does not contain
     #          any DynamicTableRegion references (i.e., we have reached the end of our table hierarchy).
@@ -75,7 +81,7 @@ def to_hierarchical_dataframe(dynamic_table):
                 #  iii) the index (i.e., id) from our target row
                 index_data = ([dynamic_table.id[row_index], ] +
                               [dynamic_table[row_index, colname]
-                               for colname in dynamic_table.colnames if colname != hcol_name])
+                               for colname in dynamic_table_colnames if colname != hcol_name])
                 index.append(tuple(index_data))
 
         # Determine the names for our index and columns of our output table
@@ -83,18 +89,15 @@ def to_hierarchical_dataframe(dynamic_table):
         # NOTE: While for a regular DynamicTable the "colnames" property will give us the full list of column names,
         #       for AlignedDynamicTable we need to use the get_colnames() function instead to make sure we include
         #       the category table columns as well.
-        iter_columns = (dynamic_table.get_colnames(include_category_tables=True, ignore_category_ids=False)
-                        if isinstance(dynamic_table, AlignedDynamicTable)
-                        else dynamic_table.colnames)
         index_names = ([(dynamic_table.name, 'id')] +
                        [(dynamic_table.name, colname)
-                        for colname in iter_columns if colname != hcol_name])
+                        for colname in dynamic_table_colnames if colname != hcol_name])
         # Determine the name of our columns
-        iter_columns = (hcol_target.get_colnames(include_category_tables=True, ignore_category_ids=False)
-                        if isinstance(hcol_target, AlignedDynamicTable)
-                        else hcol_target.colnames)
+        hcol_iter_columns = (hcol_target.get_colnames(include_category_tables=True, ignore_category_ids=False)
+                             if isinstance(hcol_target, AlignedDynamicTable)
+                             else hcol_target.colnames)
         columns = pd.MultiIndex.from_tuples([(hcol_target.name, 'id'), ] +
-                                            [(hcol_target.name, c) for c in iter_columns],
+                                            [(hcol_target.name, c) for c in hcol_iter_columns],
                                             names=('source_table', 'label'))
 
     # Case 2:  Our DynamicTableRegion columns points to another table with a DynamicTableRegion, i.e.,
@@ -120,17 +123,14 @@ def to_hierarchical_dataframe(dynamic_table):
                     # Determine the multi-index tuple for our row,
                     index_data = ([dynamic_table.id[row_index], ] +
                                   [dynamic_table[row_index, colname]
-                                   for colname in dynamic_table.colnames if colname != hcol_name] +
+                                   for colname in dynamic_table_colnames if colname != hcol_name] +
                                   list(row_tuple_level3[0]))
                     index.append(tuple(index_data))
         # Determine the names for our index and columns of our output table
         # We need to do this even if our table was empty (i.e. even is len(rows)==0)
-        iter_columns = (dynamic_table.get_colnames(include_category_tables=True, ignore_category_ids=False)
-                        if isinstance(dynamic_table, AlignedDynamicTable)
-                        else dynamic_table.colnames)
         index_names = ([(dynamic_table.name, "id")] +
                        [(dynamic_table.name, colname)
-                        for colname in iter_columns if colname != hcol_name] +
+                        for colname in dynamic_table_colnames if colname != hcol_name] +
                        hcol_hdf.index.names)
         columns = hcol_hdf.columns